Updated layouts and fixed minor bugs
