+++ app/src/main/java/vn/edu/usth/wordpressclient/MeActivity/AppSettingsActivity.java
@@ -1,9 +1,12 @@
 package vn.edu.usth.wordpressclient.MeActivity;
 
 import android.os.Bundle;
+import android.view.View;
+import android.widget.LinearLayout;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.content.ContextCompat;
 
@@ -16,8 +19,11 @@ public class AppSettingsActivity extends AppCompatActivity {
         setContentView(R.layout.activity_app_settings);
         EdgeToEdge.enable(this);
 
+        LinearLayout dark_light_btn = findViewById(R.id.appearance);
+
         Toolbar toolbar = findViewById(R.id.toolbar_app_settings);
         setSupportActionBar(toolbar);
+
         getSupportActionBar().setDisplayShowTitleEnabled(false);
         toolbar.setTitle(getString(R.string.app_settings));
 
@@ -25,5 +31,20 @@ public class AppSettingsActivity extends AppCompatActivity {
         toolbar.setNavigationOnClickListener(v -> {
             finish();
         });
+        dark_light_btn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                int currentMode = AppCompatDelegate.getDefaultNightMode();
+
+                if (currentMode == AppCompatDelegate.MODE_NIGHT_YES) {
+                    // Light
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
+                } else {
+                    // Dark
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
+                }
+            }
+        });
     }
+
 }
