+++ app/src/main/res/layout/activity_first_login.xml
@@ -66,16 +66,4 @@
         app:layout_constraintTop_toBottomOf="@+id/login_signup_btn"
         app:layout_constraintVertical_bias="0.0" />
 
-    <Button
-        android:id="@+id/dark_light_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Dark/Light"
-        android:textColor="@color/onButton"
-        app:layout_constraintBottom_toTopOf="@+id/imageView2"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.5" />
 </androidx.constraintlayout.widget.ConstraintLayout>
