+++ app/src/main/res/layout/fragment_published_post.xml
@@ -8,34 +8,6 @@
     tools:context=".PostsFragments.PublishedPostFragment">
 
 
-    <TextView
-        android:id="@+id/textView1"
-        android:textSize="16dp"
-        android:layout_width="wrap_content"
-        android:layout_height="28dp"
-        android:text="@string/you_don_t_have_any_published_post"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.496"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.549"
-        android:textColor="@color/onBackGround"/>
-
-    <Button
-        android:id="@+id/draft_post_button"
-        android:layout_width="wrap_content"
-        android:layout_height="40dp"
-        android:backgroundTint="@color/blue"
-        android:text="@string/create_a_post"
-        app:cornerRadius="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView1"
-        app:layout_constraintVertical_bias="0.028"
-        android:textColor="@color/onButton"/>
-
     <ImageView
         android:id="@+id/imageView"
         android:layout_width="173dp"
@@ -45,8 +17,34 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.391"
-        app:layout_constraintHorizontal_bias="0.5"
-        />
+        app:layout_constraintVertical_bias="0.35" />
+
+    <Button
+        android:id="@+id/draft_post_button"
+        android:layout_width="wrap_content"
+        android:layout_height="44dp"
+        android:backgroundTint="@color/blue"
+        android:text="@string/create_a_post"
+        android:textColor="@color/onButton"
+        app:cornerRadius="8dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textView1"
+        app:layout_constraintVertical_bias="0.0" />
+
+    <TextView
+        android:id="@+id/textView1"
+        android:layout_width="wrap_content"
+        android:layout_height="28dp"
+        android:text="@string/you_don_t_have_any_published_post"
+        android:textColor="@color/onBackGround"
+        android:textSize="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.516" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
